// todoList-Frontend/src/common/enum/Enum.tsx
export enum typesEnum {
  GETTODOLIST = 'GETTODOLIST',
  CREATE = 'CREATE',
  UPDATE = 'UPDATE',
  REMOVE = 'REMOVE',
  SHOWMODALUPDATE = 'SHOWMODALUPDATE',
  CLOSEMODALUPDATE = 'CLOSEMODALUPDATE',
  SHOWMODALCREATE = 'SHOWMODALCREATE',
  CLOSEMODALCREATE = 'CLOSEMODALCREATE',
  TOGGLEDARK = 'TOGGLEDARK',
}

export enum restApiEnum {
  GET = 'GET',
  POST = 'POST',
  PUT = 'PUT',
  DELETE = 'DELETE',
}